requirements.txt
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
mysql-connector-python-rf==2.0.4
oauthlib==3.2.1
requests==2.28.1